pyproject.toml
[project]
name = "Transtreaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
classifiers = ["Private :: Do Not Upload"]
requires-python = ">=3.10"
dependencies = [
  "torch>=2.5.1",
  "torchvision>=0.20.1",
  "torchaudio>=2.5.1",
  "lightning~=2.5.1.post0",
  "torchmetrics~=1.7.3",
  "scipy~=1.15.3",
  "numpy~=1.26.4",
  "kornia~=0.8.1",
  "matplotlib~=3.10.3",
  "tqdm~=4.67.1",
  "rich~=14.0.0",
  "pillow~=11.2.1",
  "jaxtyping~=0.3.2",
  "typeguard~=4.4.3",
  "grpcio~=1.73.0",
  "grpcio-tools~=1.73.0",
  "protobuf~=6.31.1",
  "hydra-core~=1.3.2",
  "hydra-colorlog~=1.2.0",
  "hydra-optuna-sweeper~=1.2.0",
  "hydra-joblib-launcher~=1.2.0",
  "omegaconf~=2.3.0",
  "optuna~=2.10.1",
  "tensorboard~=2.19.0",
  "pycocotools==2.0.7",
  "sap-toolkit~=1.0.0",
  "nuscenes-devkit~=1.1.9",
  "rootutils~=1.0.7",
  "pre-commit~=4.2.0",
  "pytest~=8.4.0",
  "sh~=2.2.2 ; sys_platform == 'linux'",
  "pywin32 ; sys_platform == 'win32'",
  "ninja>=1.11.1.4",
  "spatial-correlation-sampler>=0.5.0",
]

[project.optional-dependencies]
cpu = [
  "torch~=2.5.1",
  "torchvision~=0.20.1",
  "torchaudio~=2.5.1",
]
cu124 = [
  "torch~=2.5.1",
  "torchvision~=0.20.1",
  "torchaudio~=2.5.1",
]
cu128 = [
  "torch~=2.7.0",
  "torchvision~=0.22.0",
  "torchaudio~=2.7.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
  { index = "pytorch-cu124", extra = "cu124", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cu128", extra = "cu128", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
  { index = "pytorch-cu124", extra = "cu124", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cu128", extra = "cu128", marker = "sys_platform == 'linux'" },
]
torchaudio = [
  { index = "pytorch-cpu", extra = "cpu", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
  { index = "pytorch-cu124", extra = "cu124", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cu128", extra = "cu128", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cpu"
# url = "https://download.pytorch.org/whl/cpu"
url = "https://mirrors.nju.edu.cn/pytorch/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
# url = "https://download.pytorch.org/whl/cu124"
url = "https://mirrors.nju.edu.cn/pytorch/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
# url = "https://download.pytorch.org/whl/cu128"
url = "https://mirrors.nju.edu.cn/pytorch/whl/cu128"
explicit = true

[tool.uv]
no-build-isolation-package = ['flash-attn', 'spatial-correlation-sampler']
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu124" },
    { extra = "cu128" },
  ],
]
